==> 3. Model_averaging_and_forecasting.Rmd <==
---
title: |
  |Introduction to Species distribution modelling
  | Model forecasting and averaging
author: "Gonzalo Albaladejo-Robles"
date: "27/02/2025"
output:
  html_document: default
  pdf_document: default
---
# Project options
```{r Options, include=FALSE}
rm(list=ls())
gc()
#.rs.restartR()
options(java.parameters = "-Xmx5g") # increase the memory space for jave before loading any package
options("rgdal_show_exportToProj4_warnings"="none") # Silence packages updates warnings

setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # Set the working directory to the directory in which the code is stored
td<-tempdir()
dir.create(td,showWarnings = FALSE)
knitr::opts_chunk$set(echo = TRUE)

```

# Set up the environment
```{r Setup environment, include=FALSE}
# 0. Load the packages needed ----
list.of.packages<-c("dplyr","sf","data.table","terra","tidyr","parallel","purrr","MASS","nnet","biomod2",
                    "data.table","doParallel","rJava","rstudioapi","raptr","dismo","randomForest","caret")

new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

lapply(list.of.packages,require,character.only=TRUE)
rm(list.of.packages,new.packages)

# 0.1 Load the functions to run the analysis ----
  functions<-"./Functions" %>% list.files(recursive = FALSE,pattern = ".R$",full.names = TRUE)
  lapply(functions,function(x) source(x))

# 0.1 Load the functions to run the analysis ----
  functions<-"./Functions" %>% list.files(recursive = FALSE,pattern = ".R$",full.names = TRUE)
  lapply(functions,function(x) source(x))  

```

# 1. Load the species spatial data
```{r}
# 1. Load the species spatial information ---- 
# a. List of species----
Species_list <- list.files("./Data/SP_Info/Sp_records/Clean",full.names = TRUE)

# b. Species spatial information----
IUCN_ranges <- "./Data/Spatial/Processed/Vector/IUCN_ranges.shp" %>% st_read()
Stdy_area <- "./Data/Spatial/Processed/Vector/StdArea.shp" %>% st_read()
```

# 1.1 Load the environmental "present" information
```{r}
# c. Load the present information----
env_lyrs <- lapply("./Data/Spatial/Processed/Raster" %>% list.files(patter=".tif$",full.names = TRUE),rast)
env_data <- do.call("c",env_lyrs[-c(3)])
```

* a. Display the information
```{r }
plot(env_data[[c(1:3,21:23)]],axes=F)
```

# 1.2 Load the future scenarios of environmental change
```{r}
# d. Load the future scenarios data
terrain.lyrs <- do.call("c",env_lyrs[-c(1,3)])

fut_env <- lapply("./Data/Spatial/Processed/Scenarios" %>% list.files(patter=".tif$",full.names = TRUE,recursive = TRUE),rast)
names(fut_env)<-basename("./Data/Spatial/Processed/Scenarios" %>% list.files(patter=".tif$",full.names = TRUE,recursive = TRUE)) %>% gsub(pattern=".tif",replacement="")

fut_env <- lapply(fut_env,function(x) c(x,terrain.lyrs))
names(fut_env$Bio_Spp_126)

```
```{r}
names(fut_env$Bio_Spp_126) %in% names(env_data)
```

* b. Display the information
```{r}
panel(c(fut_env$Bio_Spp_126$wc2.1_2.5m_bio_1,
          fut_env$Bio_Spp_245$wc2.1_2.5m_bio_1,
            fut_env$Bio_Spp_370$wc2.1_2.5m_bio_1,
              fut_env$Bio_Spp_585$wc2.1_2.5m_bio_1),col=viridis::inferno(200))
```

# 2. load the saved models----
```{r}
models<-readRDS(list.files(paste("./Results/SDM_models","Eidolon helvum",sep="/"),pattern=".rds$",full.names = TRUE))
     
```

# 2.1 Predict the future scenarios GLM
```{r}
p.GLM<-predict(env_data,models$GLM,type="response") # Present scenario
f.GLM<-lapply(fut_env,function(x) predict(x,models$GLM,type="response")) # Future scenario

```

* a Compare the scenarios
```{r}
plot(p.GLM,col=viridis::inferno(100))
panel(rast(f.GLM),col=viridis::inferno(100))
```

# 2.2 Predict the future scenarios RF
```{r}

```

* a Compare the scenarios
```{r}

```

# 2.3 Predict the future scenarios ANN
```{r}

```

* a Compare the scenarios
```{r}

```

# 2.4 Predict the future scenarios MaxEnt
```{r}

```

* a Compare the scenarios
```{r}

```


PLAY WITH THE SCENARIOS!
